Session_5/pyproject.toml
[project]
name = "session-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.25.0",
    "mcp>=1.11.0",
    "rich>=14.0.0",
]
